times converting correctly and dates submitting
