+++ src/components/dashboardComponents/CreateTemplateForm.js
@@ -12,50 +12,8 @@ const addTemplate = async (data, { googleId }) => {
       `${process.env.REACT_APP_ENDPOINT_URL}/api/template`,
       template
     );
-
-    // variable makes possible manipulation of object returned from response.data
-    let newResponse;
-
-    // code converts response.data.starttime to number
-    let splitStartTime = response.data.starttime.split(':');
-    let joinStartTime = splitStartTime.join('');
-    let startTimeAsNumber = parseInt(joinStartTime, 10);
-
-    // code converts response.data.endtime to number
-    let splitEndTime = response.data.endtime.split(':');
-    let joinEndTime = splitEndTime.join('');
-    let endTimeAsNumber = parseInt(joinEndTime, 10);
-
-    // fn for converting response.data.starttime and/or endtime back to time string (from number)
-    function convertToTime(value, index) {
-      return value.substring(0, index) + ":" + value.substring(index);
-    }
-
-    // converts times from 24 hour to 12 hour format
-    if (startTimeAsNumber >= 1300) {
-      startTimeAsNumber -= 1200;
-      let startTimeAsString = startTimeAsNumber.toString();
-      console.log(`event start time: ${convertToTime(startTimeAsString, startTimeAsString.length - 2)}`);
-      let convertedStartTime = convertToTime(startTimeAsString, startTimeAsString.length - 2);
-      newResponse = { ...response.data, starttime: convertedStartTime + 'pm' };
-      if (endTimeAsNumber >= 1300) {
-        endTimeAsNumber -= 1200;
-        let endTimeAsString = endTimeAsNumber.toString();
-        let convertedEndTime = convertToTime(endTimeAsString, endTimeAsString.length - 2);
-        newResponse = { ...newResponse, endtime: convertedEndTime + 'pm' };
-      } else {
-        newResponse = { ...newResponse, endtime: newResponse.endtime + 'am' };
-      }
-    } else if (endTimeAsNumber >= 1300) {
-      endTimeAsNumber -= 1200;
-      let endTimeAsString = endTimeAsNumber.toString();
-      let convertedEndTime = convertToTime(endTimeAsString, endTimeAsString.length - 2);
-      newResponse = { ...response.data, starttime: response.data.starttime + 'am', endtime: convertedEndTime + 'pm' };
-    } else {
-      newResponse = { ...response.data, starttime: response.data.starttime + 'am', endtime: response.data.endtime + 'am' };
-    }
-    console.log(newResponse);
-    return newResponse;
+    console.log('response.data', response.data)
+    return response.data;
   } catch (error) {
     console.log(error);
   }
